AuthService/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
sqlalchemy
psycopg2-binary
